requirements.txt
streamlit
langchain-groq
langchain-core
langchain-community
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
langchain-huggingface
unstructured
